src/modified_models/ViLa-PIP/requirements.txt
# PBIP Project Requirements - Based on TPRO Environment
# Generated from actual working environment

# Deep Learning Framework
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Computer Vision and Image Processing
opencv-python==4.9.0.80
albumentations==1.4.2
scikit-image==0.22.0
Pillow==10.2.0

# Scientific Computing
numpy==1.23.5
scipy==1.12.0
scikit-learn==1.4.1.post1

# Data Visualization
matplotlib==3.8.4

# Utilities
tqdm==4.66.5
omegaconf==2.3.0
packaging>=23.0  # For version checking

# Test-Time Augmentation
ttach==0.0.3

# Medical Image Processing
MedCLIP==0.0.3

# Data Processing
pandas==2.2.1

# Additional Dependencies
PyYAML==6.0.1
requests==2.32.3

# Optional but recommended for enhanced functionality
timm==0.9.16              # For additional model architectures
transformers==4.24.0      # For transformer models
tokenizers==0.13.3        # For text processing
huggingface-hub==0.26.1   # For model downloads

# Development and Analysis Tools (optional)
jupyter==1.0.0
jupyterlab==4.0.11
tensorboardX==2.6.2.2     # For logging and visualization 